# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GST Billing & Inventory Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      security: []  # public: no session required
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change your own password
      description: >-
        Requires the current password. On success the session cookie is
        re-issued, so the browser is not left holding one minted under the old
        credential.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                required: [success]
        "400":
          description: The new password does not meet the policy
        "403":
          description: The current password is incorrect
  /auth/register:
    post:
      operationId: register
      security: []  # public: no session required
      tags: [auth]
      summary: Register new business user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # Vendors
...
security:
  # Every operation requires a session unless it overrides this with
  # `security: []`. The spec previously declared no authentication at all, so
  # the generated clients — and anyone reading it — saw a wholly public API.
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gst_session
      description: >-
        Session cookie issued by /auth/login and /auth/register. HttpOnly, so
        page scripts cannot read it. Browser clients must also send the CSRF
        token from the readable `gst_csrf` cookie in an `X-CSRF-Token` header
        on every state-changing request.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        For native clients that have no cookie jar. Browser clients should use
        the cookie instead, so the token is never reachable from script.
  schemas:
...
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
          description: At least 12 characters; screened against common and breached passwords.
      required: [currentPassword, newPassword]
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password, businessName]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        businessName:
          type: string
        gstin:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        subscriptionStatus:
          type: ["string", "null"]
        subscriptionEnd:
          type: ["string", "null"]
        businessId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Customer:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        gstin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        stateCode:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        creditLimit:
          type: ["number", "null"]
        outstandingBalance:
          type: ["number", "null"]
        createdAt:
          type: string
    CustomerListResponse:
      type: object
      required: [customers, total]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        gstin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        stateCode:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        creditLimit:
          type: ["number", "null"]
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        gstin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        stateCode:
          type: ["string", "null"]
        pincode:
          type: ["string", "null"]
        creditLimit:
          type: ["number", "null"]
...